backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.6.1
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whisper-openai==20231117
instructor==0.4.8
django-cors-headers==4.3.1
celery==5.3.6
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
prometheus-client==0.19.0
kubernetes==29.0.0
boto3==1.34.29
azure-mgmt-compute==30.3.0
google-cloud-compute==1.15.0 